requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
pydantic==2.12.3
python-multipart==0.0.6
PyMuPDF==1.22.6
pdf2image==1.17.0
pillow==12.0.0
pytesseract==0.3.10
paddleocr==3.3.1
paddlepaddle==3.2.1
requests==2.32.5
transformers==4.57.1
torch>=2.0.0
scikit-learn
spacy
matplotlib
shap
lime
pytest
